chore: code review I
